flags

+++ philo/atoi_spe.c
@@ -6,13 +6,13 @@
 /*   By: tchevrie <tchevrie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/29 16:31:56 by tchevrie          #+#    #+#             */
-/*   Updated: 2023/03/29 16:41:19 by tchevrie         ###   ########.fr       */
+/*   Updated: 2023/04/03 14:27:10 by tchevrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-static long	_bis(const char *str, int i, long nbr, long sign)
+static long long	_bis(const char *str, int i, long long nbr, long long sign)
 {
 	while ('0' <= str[i] && str[i] <= '9')
 	{
@@ -26,10 +26,10 @@ static long	_bis(const char *str, int i, long nbr, long sign)
 	return (nbr * sign);
 }
 
-long	atoi_spe(const char *str)
+long long	atoi_spe(const char *str)
 {
-	long		nbr;
-	long		sign;
+	long long		nbr;
+	long long		sign;
 	size_t		i;
 
 	nbr = 0;
